also had to change the return type for pp calc

+++ ScoreCalc/ScoreCalculator.cs
@@ -29,7 +29,7 @@ namespace osuscorefetcher.ScoreCalc
         /// </summary>
         /// <param name="score">Score data from the API</param>
         /// <returns>Score's PP</returns>
-        public async Task<double?> CalculateScorePPAsync(ApiClasses.Score score)
+        public async Task<float?> CalculateScorePPAsync(ApiClasses.Score score)
         {
             // preparing necessary data
             Ruleset ruleset = GetRulesetFromScore(score);
@@ -52,7 +52,7 @@ namespace osuscorefetcher.ScoreCalc
             PerformanceCalculator performanceCalculator = ruleset.CreatePerformanceCalculator();
             PerformanceAttributes performanceAttributes = await performanceCalculator.CalculateAsync(scoreInfo, difficultyAttributes, default);
 
-            return performanceAttributes.Total;
+            return (float)performanceAttributes.Total;
         }
         /// <summary>
         /// Prepare ScoreInfo object for use in calculating difficulty and performance attributes
